requirements.txt
tensorflow-cpu
streamlit==0.78.0
tensorflow_hub==0.11.0
numpy==1.19.5
matplotlib==3.2.2
Pillow==8.1.2
